Split out verification. Fixed identity_document
