+++ identity_document/src/diff/mod.rs
@@ -2,14 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub use self::diff_document::DiffDocument;
-pub use self::diff_method::DiffMethod;
+pub use identity_verification::verification_method::diff::DiffMethod;
 pub use self::diff_service::DiffService;
-pub use self::method_data::DiffMethodData;
-pub use self::method_ref::DiffMethodRef;
+pub use identity_verification::verification_method::diff::DiffMethodData;
+pub use identity_verification::verification_method::diff::DiffMethodRef;
 
 mod diff_document;
-mod diff_method;
 mod diff_service;
-mod method_data;
-mod method_ref;
-mod method_type;
