+++ identity_verification/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "identity_verification"
+version = "0.7.0-alpha.4"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+identity_core = { version = "=0.7.0-alpha.4", path = "./../identity_core", default-features = false }
+identity_did = { version = "=0.7.0-alpha.4", path = "./../identity_did", default-features = false }
+serde.workspace = true
+strum.workspace = true
+thiserror.workspace = true
+
+[dev-dependencies]
+serde_json = { version = "1.0", default-features = false }
