+++ identity_verification/src/verification_method/diff/mod.rs
@@ -1,0 +1,11 @@
+// Copyright 2020-2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod diff_method;
+mod method_data;
+mod method_ref;
+mod method_type;
+
+pub use self::diff_method::DiffMethod;
+pub use self::method_data::DiffMethodData;
+pub use self::method_ref::DiffMethodRef;
