+++ identity_verification/src/verification_method/mod.rs
@@ -7,20 +7,22 @@
 //! `identity_iota_core_legacy` crate.
 
 mod builder;
-mod method_data;
+mod material;
+mod method;
 mod method_ref;
 mod method_relationship;
 mod method_scope;
 mod method_type;
+#[deprecated(since = "0.5.0", note = "diff chain features are slated for removal")]
+pub mod diff; 
 mod traits;
-mod verification_method;
 
 pub use self::builder::MethodBuilder;
-pub use self::method_data::MethodData;
+pub use self::material::MethodData;
+pub use self::method::VerificationMethod;
 pub use self::method_ref::MethodRef;
 pub use self::method_relationship::MethodRelationship;
 pub use self::method_scope::MethodScope;
 pub use self::method_type::MethodType;
 pub use self::traits::MethodUriType;
 pub use self::traits::TryMethod;
-pub use self::verification_method::VerificationMethod;
